Integrasi data real: rpt_trx ETL + model 18 lantai + adapter DB
- Model gedung 18 lantai (B2,B1,LT1-LT16); demo & seed di-regenerate
- db/etl-rpt-trx.sql: ETL rpt_trx -> taps/employees (trxdate->WIB, extsysid=NIP,
  fname=nama, cardno=kartu, filter Valid Credential, JOIN stair_reader)
- db/stair-reader.sql: 18 reader "... Tangga Darurat Tengah" -> lantai
- src/lib/db.ts + getScores async: baca Postgres bila DATA_SOURCE=db, else JSON
- card BIGINT, README mode data real

+++ scripts/generate-sql.mjs
@@ -33,7 +33,7 @@ DROP TABLE IF EXISTS employees CASCADE;
 CREATE TABLE employees (
   id         TEXT PRIMARY KEY,
   name       TEXT NOT NULL,
-  card       INTEGER,
+  card       BIGINT,
   unit       TEXT,
   persona    TEXT,            -- kebiasaan: champion|regular|occasional|rare
   office     TEXT,            -- lantai kantor (mis. LT12)
